+++ src/shared/getMetaInfo.js
@@ -1,6 +1,6 @@
+import { ensureIsArray } from '../utils/ensure'
 import { applyTemplate } from './template'
 import { defaultInfo, disableOptionKeys } from './constants'
-import { ensureIsArray } from './ensure'
 import { escape } from './escaping'
 import getComponentOption from './getComponentOption'
