+++ src/utils/is-type.js
